Consolidate article schemas and add body field to article interface
