+++ src/articles/interfaces/article-row.interface.ts
@@ -1,7 +1,7 @@
 import type { Profile } from '@/profiles/profiles.schema';
 import type { ArticleTag } from '@/tags/tags.schema';
 import type { InferSelectModel } from 'drizzle-orm';
-import type { articles, favoriteArticles } from '../schema';
+import type { articles, favoriteArticles } from '../articles.schema';
 
 type ArticleFavoritedBy = InferSelectModel<typeof favoriteArticles>;
