+++ src/articles/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './article-row.interface';
+export * from './article.interface';
