+++ src/articles/mappers/articles.mapper.ts
@@ -1,11 +1,10 @@
 import type { ArticleResponseDto } from '../dto/article-response.dto';
-import type { ArticleRow } from '../interfaces/article-row.interface';
-import type { IArticle } from '../interfaces/article.interface';
+import type { Article, ArticleRow } from '../interfaces';
 
 export function toDomain(
   article: ArticleRow,
   { currentUserId }: { currentUserId: number | null },
-): IArticle {
+): Article {
   return {
     id: article.id, // TODO: Do we need this?
     slug: article.slug,
@@ -18,6 +17,7 @@ export function toDomain(
       (user) => user.userId === currentUserId,
     ),
     favoritesCount: article.favoritedBy.length,
+    body: article.body,
     author: {
       username: article.author.username,
       bio: article.author.bio,
@@ -29,7 +29,7 @@ export function toDomain(
   };
 }
 
-export function toResponse(article: IArticle): ArticleResponseDto {
+export function toResponse(article: Article): ArticleResponseDto {
   return {
     slug: article.slug,
     title: article.title,
@@ -43,7 +43,7 @@ export function toResponse(article: IArticle): ArticleResponseDto {
       image: article.author.image,
       following: false, // TODO
     },
-    body: 'TODO',
+    body: article.body,
     createdAt: article.createdAt.toISOString(),
     updatedAt: article.updatedAt.toISOString(),
   };
